feat(bootstrap): add alias scratchpad for  convenience
This commit introduces a new file, `003_scratchpad.bootstrap.zsh`, which contains a function `gch` for quickly committing changes with a provided message. Additionally, it sets up an alias `gg` for this function. The existing `001_aliases.bootstrap.zsh` file is updated to unalias `gg` if it already exists, ensuring no conflicts arise from the new alias.